# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flyspray",
        "version": "1.0.0",
        "contact": {}
    },
    "paths": {
...
        "/member": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new member for a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Create a new member",
                "parameters": [
                    {
                        "description": "Member data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.createMemberInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "member created successfully (Member details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Failed to read member data",
                        "schema": {}
                    },
                    "401": {
                        "description": "Authentication is required",
                        "schema": {}
                    },
                    "403": {
                        "description": "Access denied to create member",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/member/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the ownership of a member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Update member ownership",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Member data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.updateMemberInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member ownership updated successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Failed to read member data",
                        "schema": {}
                    },
                    "401": {
                        "description": "Authentication is required",
                        "schema": {}
                    },
                    "403": {
                        "description": "Access denied to update member",
                        "schema": {}
                    },
                    "404": {
                        "description": "Member is not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/member/{project_id}": {
            "get": {
                "description": "Get a list of members in a project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Get members in a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "members in project retrieved successfully (Members details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.ResponseMsg": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.createMemberInput": {
            "type": "object",
            "required": [
                "project_id",
                "user_id"
            ],
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "project_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "app.updateMemberInput": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "boolean"
                },
                "project_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}